XW-849 rename auth
