+++ front/auth/app/common/auth-logger.js
