+++ front/auth/app/common/visit-source-wrapper.js
