+++ front/auth/app/component/auto-tab.js
