+++ front/auth/app/facebook/facebook-registration.js
