+++ front/auth/tests/unit/common/Cookie.js
@@ -1,4 +1,4 @@
-QUnit.module('auth/common/Cookie', function (hooks) {
+QUnit.module('auth/common/cookie', function (hooks) {
 	/**
 	 * TODO: replace this with a class method to delete a cookie
 	 * @param {string} name
@@ -21,11 +21,11 @@ QUnit.module('auth/common/Cookie', function (hooks) {
 	});
 
 	QUnit.test('Cookie class is loaded', function (assert) {
-		assert.ok(typeof require('auth/app/common/Cookie').default === 'function');
+		assert.ok(typeof require('auth/app/common/cookie').default === 'function');
 	});
 
 	QUnit.test('Cookie getter is extracting right value from a cookie', function (assert) {
-		var Cookie = require('auth/app/common/Cookie').default;
+		var Cookie = require('auth/app/common/cookie').default;
 
 		assert.equal(Cookie.get('Geo'), '{"city":"FIXME","country":"PL","continent":"EU"}');
 		assert.equal(Cookie.get('i18next'), 'en');
@@ -33,13 +33,13 @@ QUnit.module('auth/common/Cookie', function (hooks) {
 	});
 
 	QUnit.test('Cookie getter is returning null if cookie value not found', function (assert) {
-		var Cookie = require('auth/app/common/Cookie').default;
+		var Cookie = require('auth/app/common/cookie').default;
 
 		assert.equal(Cookie.get('fakeCookieValue'), null);
 	});
 
 	QUnit.test('Cookie getter is extracting right value from the end of cookie string', function (assert) {
-		var Cookie = require('auth/app/common/Cookie').default;
+		var Cookie = require('auth/app/common/cookie').default;
 
 		assert.equal(Cookie.get('i18next'), 'en');
 	});
