+++ front/auth/tests/unit/common/Geo.js
@@ -2,7 +2,7 @@ QUnit.module('auth/common/Geo)', function (hooks) {
 	hooks.beforeEach(function () {
 		var geoModule = {};
 
-		require.entries['auth/app/common/Geo'].callback(geoModule, {
+		require.entries['auth/app/common/geo'].callback(geoModule, {
 			get: function () {
 				return '{"city":"FIXME","country":"PL","continent":"EU"}';
 			}
