+++ front/auth/tests/unit/common/form-errors.js
@@ -1,4 +1,4 @@
-QUnit.module('auth/common/FormErrors)', function (hooks) {
+QUnit.module('auth/common/form-errors)', function (hooks) {
 	hooks.beforeEach(function () {
 		var form = document.createElement('form'),
 			AuthTracker = function () {
@@ -12,7 +12,7 @@ QUnit.module('auth/common/FormErrors)', function (hooks) {
 			},
 			FormModule = {};
 
-		require.entries['auth/app/common/FormErrors'].callback(FormModule, AuthTracker, null);
+		require.entries['auth/app/common/form-errors'].callback(FormModule, AuthTracker, null);
 
 		this.formErrors = new FormModule.default(form);
 
@@ -30,7 +30,7 @@ QUnit.module('auth/common/FormErrors)', function (hooks) {
 	});
 
 	QUnit.test('FormErrors class is loaded', function (assert) {
-		assert.ok(typeof require('auth/app/common/FormErrors').default === 'function');
+		assert.ok(typeof require('auth/app/common/form-errors').default === 'function');
 	});
 
 	QUnit.test('Displaying field error', function (assert) {
