+++ server/tests/unit/facets/operations/prepare-curated-content-data.js
@@ -1,4 +1,4 @@
-QUnit.module('facets/operations/prepareCuratedContentData');
+QUnit.module('facets/operations/prepare-curated-content-data');
 
 QUnit.test('getTitle', function (assert) {
 	var testCases = [
